requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
attrs==24.2.0
Babel==2.13.1
bcrypt==4.0.1
billiard==4.2.2
Brotli==1.1.0
celery==5.5.3
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
crispy-tailwind==1.0.3
cryptography==41.0.7
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.10
django-allauth==0.57.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-extensions==4.1
django-filter==23.3
django-redis==6.0.0
django-timezone-field==7.1
django-user-agents==0.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
et_xmlfile==2.0.0
fonttools==4.59.2
frozenlist==1.8.0
geoip2==5.1.0
gunicorn==21.2.0
html5lib==1.1
idna==3.10
importlib_resources==6.5.2
joblib==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
kombu==5.5.4
matplotlib==3.7.2
maxminddb==2.8.2
multidict==6.7.0
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyparsing==3.0.9
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
qrcode==8.2
redis==6.4.0
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.18.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
sentry-sdk==1.38.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
statsmodels==0.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.6.0
XlsxWriter==3.1.9
xmltodict==0.13.0
yarl==1.20.1
zipp==3.23.0
zopfli==0.2.3.post1
